move all tools to a folder
