+++ app/osint/tools/_numverify.py
@@ -1,0 +1,36 @@
+import requests
+from ... import config
+from ..utils import (
+    failed_to_run,
+    missing_apikey,
+)
+
+
+def numverify(indicator):
+    try:
+        if config["NUMVERIFY_API_KEY"] == "":
+            return missing_apikey("numverify")
+
+        response = requests.get(
+            f"http://apilayer.net/api/validate?access_key={config['NUMVERIFY_API_KEY']}&number={indicator.indicator}&format=1"
+        )
+
+        if response.status_code != 200:
+            return failed_to_run(
+                tool_name="numverify",
+                status_code=response.status_code,
+                reason=response.reason,
+            )
+
+        return (
+            # fmt: off
+                {
+                    "tool": "numverify",
+                    "outcome": {"status": "results_found", "error_message": None, "status_code": response.status_code, "reason": response.reason},
+                    "results": response.json()
+                },
+            # fmt: on
+        )
+
+    except Exception as error_message:
+        return failed_to_run(tool_name="numverify", error_message=error_message)
